feat(roach): upgrade roach movement/rotation
Roaches can now move on complex geometry without clipping through walls. This was done
by rotating the roach based on the surface normal of the surface under them.

I need to make the rotation and position changes smooth between corner traversals and this will be looking niiiiiiice.

I had issues with the roaches clipping through walls and then not being able to traverse a corner. To fix this,
if the roach can't find a spot to move to on a corner traversal, it will fall to the ground. I haven't let a bunch of them roam for a super long time yet,
but I'm pretty sure that the roaches will not move through walls anymore. :)