feat(character/motioncontroller): various changes
Added crouch mechanics so that the character's collision capsule updates it's height based on the camera distance from the floor.
(I need to add "floor" tags to all of the floor objects so that if the camera goes above another object, the capsule height doesn't change.

I also added haptic feedback effects to the hand controllers to let the player know when they can climb up an object.