Add stage functionality and on/off for lights
Stages:
I added some stage functionality to a stage system, where stages are objects
that change the world (spawn triggers, move props, ect...) and have a set of
flags that the player needs to complete to advance to the next stage.

This is kept track by a vector of function pointers that return a bool. If the
function returns true, then that function pointer is removed from the vector.
When the vector is empty, the stage is marked as completed, and when the player
goes through another loop, the stage will change to the next stage. (there is
lots of room for changes in funtionality)

Lights:
I wrote a quick piece of code to dynamically update the material on a light,
so as stages change, the lights can be turned on/off and the material will change
as well. (this is mainly for when you turn a light on and the light source material
changes as well so it looks visually pleasing and not wonky)