feat(light manager / grabbable):
Created a light manager class to hold pointers to all of the light instances so
they can be easily be edited (turned on/off) by stages.

Added Grabbable class to be a base class for all objects that can be grabbed to
control things such as gripsize (how much to close hand based on size of object
it is gripping)

I also added ErrolCharacter for the enemy AI so that it will be able to follow a nav
mesh.

various other small refactorings...