feat(bottle): add destructible bottle object
ErrolCharacter will react to this bottle being thrown and breaking.

I have protoyped a blueprint callable function that will notify ErrolCharacter
when the bottle breaks so he may hear it and go to investigate where the noise was.
ErrolCharacter will move to a location with a distance from the breaking point depending
on how well he could hear the break. If ErrolCharacter is farther away from the break, then
he will move to a point further away from the breaking point.

I will also have to implement funcionality to where if the bottle is moving and is in
ErrolCharacter's view, then ErrolCharacter will move to a place in the direction
of where the bottle is moving from.