feat(bottle): implement break velocity threshold
The bottle breaks if it collides and its velocity is over a certain threshold.

I was going to add mechanics in the bottle class to make it only shatter if it
was moving over a certain velocity. But after researaching the UDestructibleComponent
and UDestructubleMesh classes, there seemed to be no way to enable / disable invulnerability.

I then looked at the destructible mesh in editor and found "Impact Velocity Threshold" in the advanced tab!!!
It's just what I needed and it works perfectly now.