feat(ErrolCharacter): more work on peeking
Errol is now positioned correctly on the z axis and rotates
to face the player camera to ensure the peek looks correct.

Now I need to make errol move left and right on the right vector axis
so that not too much of him is visible when the player sees the peek.
How will I do this? Only time will tell...