feat(drawer): add fully open/closed mechanics
the drawer no longer moves if the hand controller is too far away and the drawer
is fully open or closed.

This makes the drawer feel better because if you fully close (or open) it, and you push your hand
controller into the drawer, it won't open right away if you move your hand back a bit, and it will
only open when you are back at where your hand was when it fully closed it.