work on various features:
CrawlSpaceScare:
work on crawling man's IK

Dog:
Added failsafes to the pathfinding, so if it can't
fetch or return the ball, differen't actions occur to preserve
the fetch loop.

DogController:
This was created to make the pathfinding failsafes while fetching the ball.

ErrolCharacter:
Move speed is drastically reduced when the player is using a door and Errol is not in view.