work on various features:
AudioDistanceCrossfader:
I created this actor to crossfade ambience sounds over
a distance. I currently use this to fade between the
rain/house ambience to the crawlspace ambience and vice versa.

Door:
I changed the door mechanics so that the door rotation lerps
to where the hand controller is. This gives it a smoother feel and
makes the door feel like it has weight. I figured out that Paranormal Activy VR
does this so I implemented my own version.

I also changed how the capsule collision works. It is better now. I need to
update the hand controller collision mechannics to be like this.

VRCharacter:
I update the footstep audio system a bit. This probably needs more tweaking, but it might be good.

I also optimized the GPU performance a bit.