work on StageManager and ErrolCharacter
StageManager:

Stages now advance when the current stage is nullptr.
This is a nice change that allows easier implementation of
stage features that doesn't require a stage to be made.

ErrolCharacter:

I added a function that returns true if the player can see the
mocap mesh. This will be used to determine when we can disappear Errol.