work on player death system
- The player now respawns after they are killed and the world is reset.

I have a weird bug where the player is able to walk through doors after they are killed.

I tracked down the cause of the bug: When the blood stains are spawned
when Errol is chainsawwing the player, the player collision changes.
I'm still not sure why this is happening. I will figure this out tomorrow.